+++ src/main/java/org/revo/registration/repository/CourseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 /*
  *  Created by revo (ashraf1abdelrasool@gmail.com)  on 8/31/19, 5:40 PM.
  */
-public interface CourseRepository extends CrudRepository<Course, Long> {
+public interface CourseRepository extends CrudRepository<Course, Integer> {
 }
