+++ src/main/java/org/revo/registration/repository/SCRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 /*
  *  Created by revo (ashraf1abdelrasool@gmail.com)  on 8/31/19, 5:42 PM.
  */
-public interface SCRepository extends CrudRepository<SC, Long> {
+public interface SCRepository extends CrudRepository<SC, Integer> {
 }
